refactor: LottoGenerator static 클래스로 변경
