+++ src/main/java/lotto/common/LottoGenerator.java
@@ -1,4 +1,4 @@
-package lotto.models;
+package lotto.common;
 
 import static lotto.common.LottoConfiguration.LOTTO_COUNT;
 import static lotto.common.LottoConfiguration.MAX_VALUE;
@@ -9,18 +9,15 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import lotto.models.Lotto;
 
 public class LottoGenerator {
-    private final List<Integer> lottoPreset;
+    static private final List<Integer> lottoPreset = IntStream
+            .rangeClosed(MIN_VALUE, MAX_VALUE)
+            .boxed()
+            .collect(Collectors.toList());
 
-    public LottoGenerator() {
-        lottoPreset = IntStream
-                .rangeClosed(MIN_VALUE, MAX_VALUE)
-                .boxed()
-                .collect(Collectors.toList());
-    }
-
-    public Lotto createLotto() {
+    public static Lotto createLotto() {
         Collections.shuffle(lottoPreset);
         List<Integer> numbers = lottoPreset.subList(0, LOTTO_COUNT).stream()
                 .sorted()
@@ -28,7 +25,7 @@ public class LottoGenerator {
         return new Lotto(numbers);
     }
 
-    public List<Lotto> createLottos(int numberOfLotto) {
+    public static List<Lotto> createLottos(int numberOfLotto) {
         List<Lotto> list = new ArrayList<>();
         for (int i = 0; i < numberOfLotto; i++) {
             list.add(createLotto());
